API/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bidict==0.23.1
click==8.1.7
fastapi==0.115.0
h11==0.14.0
idna==3.10
numpy==2.1.1
opencv-python==4.10.0.84
pydantic==2.9.2
pydantic_core==2.23.4
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
websockets==13.1
wsproto==1.2.0
